messaging_app/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-filter==25.1
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
flake8==7.3.0
idna==3.10
mccabe==0.7.0
parameterized==0.9.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
